chore: move shared dev config into workspace package
Consolidates the shared eslint and tsconfig bases into a new
`dev-config` workspace package and removes their top-level copies.

- packages/dev-config/eslint.base.js — was eslint.config.base.js
- packages/dev-config/tsconfig.base.json — was tsconfig.base.json
- packages/dev-config/package.json exports them under
  `./eslint` and `./tsconfig`

Each package now depends on `dev-config: workspace:*` and references
the shared bases via package-relative paths instead of
`../../<file>`:

  import base from "dev-config/eslint";
  { "extends": "dev-config/tsconfig" }

Motivation: VS Code's ESLint extension was reporting
"multiple candidate TSConfigRootDirs are present" — typescript-eslint v8
walks the V8 call stack for files matching `eslint.config.*` and adds
each as a candidate root. Having a workspace-root config encouraged that
ambiguity. With the shared base now inside a package and only
per-package `eslint.config.js` at top level, only one config sits in
the lookup path during a given lint context.
