+++ packages/agent-runtime-api/package.json
@@ -16,6 +16,7 @@
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
+    "dev-config": "workspace:*",
     "typescript": "^6.0.2"
   }
 }
