+++ packages/agent-runtime/tsconfig.json
@@ -1,4 +1,4 @@
 {
-  "extends": "../../tsconfig.base.json",
+  "extends": "dev-config/tsconfig",
   "include": ["src"]
 }
