+++ packages/api-server/package.json
@@ -44,6 +44,7 @@
   "devDependencies": {
     "@types/js-yaml": "^4.0.9",
     "@types/ws": "^8.18.1",
+    "dev-config": "workspace:*",
     "grpc-tools": "^1.13.1",
     "ts-proto": "^2.11.6",
     "tsup": "^8.5.1",
