+++ packages/cli/package.json
@@ -40,6 +40,7 @@
     "@types/ws": "^8.18.1",
     "api-server-api": "workspace:*",
     "commander": "^14.0.3",
+    "dev-config": "workspace:*",
     "open": "^11.0.0",
     "smol-toml": "^1.6.1",
     "tar-stream": "^3.1.8",
