+++ packages/db/package.json
@@ -15,6 +15,7 @@
   },
   "devDependencies": {
     "drizzle-kit": "^0.31.4",
+    "dev-config": "workspace:*",
     "typescript": "^6.0.2"
   }
 }
