+++ packages/ui/package.json
@@ -47,6 +47,7 @@
     "@types/react": "^18.3.1",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.4",
+    "dev-config": "workspace:*",
     "playwright": "^1.59.0",
     "sharp": "^0.34.5",
     "typescript": "^5.6.3",
